Refactored Terraform code: main.tf
In the refactored code, we defined the user data script as a variable to avoid repeating the same code block twice. We also used the count parameter to create two instances of the same resource with the same configuration. This makes the code more concise and easier to maintain. we added a tags block to the aws_instance resource and used the count.index interpolation syntax to append the index to the end of the resource name. The count.index starts from 0, so we added 1 to it to get the desired numbering (_1 and _2).